Move uploadFileOnVisual to visual module.
