+++ src/domain/storage/storage-bucket/storage.bucket.resolver.mutations.ts
@@ -8,14 +8,7 @@ import { AuthorizationPrivilege } from '@common/enums/authorization.privilege';
 import { StorageBucketService } from './storage.bucket.service';
 import { DocumentAuthorizationService } from '../document/document.service.authorization';
 import { FileUpload, GraphQLUpload } from 'graphql-upload';
-import { IVisual } from '@domain/common/visual/visual.interface';
-import { VisualUploadImageInput } from '@domain/common/visual/dto/visual.dto.upload.image';
-import { VisualService } from '@domain/common/visual/visual.service';
 import { DocumentService } from '../document/document.service';
-import { UpdateVisualInput } from '@domain/common/visual/dto/visual.dto.update';
-import { Visual } from '@domain/common/visual';
-import { EntityNotInitializedException } from '@common/exceptions';
-import { LogContext } from '@common/enums';
 import { StorageBucketUploadFileInput } from './dto/storage.bucket.dto.upload.file';
 import { IStorageBucket } from './storage.bucket.interface';
 import { DeleteStorageBuckeetInput as DeleteStorageBucketInput } from './dto/storage.bucket.dto.delete';
@@ -26,77 +19,8 @@ export class StorageBucketResolverMutations {
     private authorizationService: AuthorizationService,
     private storageBucketService: StorageBucketService,
     private documentAuthorizationService: DocumentAuthorizationService,
-    private documentService: DocumentService,
-    private visualService: VisualService
+    private documentService: DocumentService
   ) {}
-
-  @UseGuards(GraphqlGuard)
-  @Mutation(() => IVisual, {
-    description: 'Uploads and sets an image for the specified Visual.',
-  })
-  async uploadImageOnVisual(
-    @CurrentUser() agentInfo: AgentInfo,
-    @Args('uploadData') uploadData: VisualUploadImageInput,
-    @Args({ name: 'file', type: () => GraphQLUpload })
-    { createReadStream, filename, mimetype }: FileUpload
-  ): Promise<IVisual> {
-    const visual = await this.visualService.getVisualOrFail(
-      uploadData.visualID,
-      {
-        relations: {
-          profile: {
-            storageBucket: {
-              authorization: true,
-            },
-          },
-        },
-      }
-    );
-    this.authorizationService.grantAccessOrFail(
-      agentInfo,
-      visual.authorization,
-      AuthorizationPrivilege.UPDATE,
-      `visual image upload: ${visual.id}`
-    );
-    const profile = (visual as Visual).profile;
-    if (!profile || !profile.storageBucket)
-      throw new EntityNotInitializedException(
-        `Unable to find profile or storageBucket for Visual: ${visual.id}`,
-        LogContext.STORAGE_BUCKET
-      );
-
-    const storageBucket = profile.storageBucket;
-    // Also check that the acting agent is allowed to upload
-    this.authorizationService.grantAccessOrFail(
-      agentInfo,
-      storageBucket.authorization,
-      AuthorizationPrivilege.FILE_UPLOAD,
-      `visual image upload on storage bucket: ${visual.id}`
-    );
-    const readStream = createReadStream();
-
-    const visualDocument = await this.storageBucketService.uploadImageOnVisual(
-      visual,
-      storageBucket,
-      readStream,
-      filename,
-      mimetype,
-      agentInfo.userID
-    );
-
-    // Ensure authorization is updated
-    await this.documentAuthorizationService.applyAuthorizationPolicy(
-      visualDocument,
-      storageBucket.authorization
-    );
-    const updateData: UpdateVisualInput = {
-      visualID: visual.id,
-      uri: this.documentService.getPubliclyAccessibleURL(visualDocument),
-      alternativeText: uploadData.alternativeText,
-    };
-    return await this.visualService.updateVisual(updateData);
-  }
-
   @UseGuards(GraphqlGuard)
   @Mutation(() => String, {
     description:
